Added new database folder for better separation and ongoing migration updates
A creation of a new database folder, which separates the API routes calls in a more organized manner. Additionally, there are ongoing updates for the migration to a new structure in users and organizations. During testing, some bugs were discovered in the unitary function, which require fixing. Some of these bugs have already been resolved, and the remaining ones will be fixed soon.