Debugging Transaction List Data Retrieval Issue
Work on debugging a transaction list data retrieval issue. Currently, there is a bug causing the data to not be properly retrieved from the database. However, despite this issue, the main data is being correctly retrieved and the front-end is responding as expected. Additional work is being done to resolve the transaction list data issue.