Organization Selection on Login
This commit addresses an issue that was discovered during testing with users. Previously, there was no way for users to select an organization during login, which caused confusion and made it difficult to access the correct data. To fix this,  added an organization selection feature to the sign-in page, which allows users to choose the correct organization before proceeding. In addition, work is ongoing to ensure that this selection is saved to local storage, which will make the process even smoother for users.