Alpha Release
This commit includes optimizations and fixes to prepare for a full production deploy of the alpha version. A build run has been applied to ensure the app is optimized for the release. Bugs have been fixed and improvements made in preparation for the upcoming alpha version, which will be ready for testing in a production environment.