Feature small UI upgrade (#4) (#5)
* Fix: Helper button and Google sign-in visibility

Two errors were fixed in this commit. Firstly, a helper develop button was mistakenly left in the main section and  has been removed from the app. Secondly, the Google sign-in was left hidden by mistake, and has been fixed and the Google sign-in is now visible to users.

* Minor changes

The base button was not commented on correctly. Fixed