Merge all commits (#6)
* UI improvements and code optimization

This commit includes several UI improvements and code optimizations. The CurrentBudget component was removed from the main view in the Home page as it was not yet ready for release. Additionally, several React hook dependencies were ignored using "ignore next ESLint line" to improve code readability. Unused console.logs were also removed to improve code efficiency. Finally, the navbar resizing issue was fixed to ensure consistent sizing in all states.

* Added descriptive README file to root directory

This commit adds a detailed README file to the root directory of the project, providing information about the purpose of the application, how to install and run it, and other relevant information. The README is designed to give users a better understanding of the app and its features.

* Fix: Helper button and Google sign in visibility

Two errors were fixed in this commit. Firstly, a helper develop button was mistakenly left in the main section and  has been removed from the app. Secondly, the Google sign in was left hidden by mistake, and has been fixed and the Google sign in is now visible for users.

* Minor changes

The base button was not commented on correctly. Fixed

* Feature small UI upgrade (#4) (#5)

* Fix: Helper button and Google sign-in visibility

Two errors were fixed in this commit. Firstly, a helper develop button was mistakenly left in the main section and  has been removed from the app. Secondly, the Google sign-in was left hidden by mistake, and has been fixed and the Google sign-in is now visible to users.

* Minor changes

The base button was not commented on correctly. Fixed